7039 Move more tests to InstanceConfigurationScreenMockTest
